+++ get_next_line_utils.c
@@ -5,24 +5,21 @@
 /*                                                    +:+ +:+         +:+     */
 /*   By: fschuber <fschuber@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/10/20 06:43:51 by fschuber          #+#    #+#             */
-/*   Updated: 2023/10/20 06:52:26 by fschuber         ###   ########.fr       */
+/*   Created: 2023/10/23 07:07:11 by fschuber          #+#    #+#             */
+/*   Updated: 2023/10/23 08:13:02 by fschuber         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "get_next_line.h"
 
 /*
-	@return			Substring
-
-	@param s		Input String
-	@param start	Substring Start Index
-	@param len		Substring length
+	@brief Returns a substring of string s of length len beginning at index start
 */
 char	*gnl_substr(char const *s, unsigned int start, size_t len)
 {
 	char		*p;
 	int			substr_len;
+	size_t		counter;
 
 	substr_len = len;
 	if (ft_strlen(s) - start < len)
@@ -34,20 +31,44 @@ char	*gnl_substr(char const *s, unsigned int start, size_t len)
 	p = (char *)malloc(substr_len + 1);
 	if (p == NULL)
 		return (NULL);
-	ft_memcpy(p, &s[start], substr_len);
+	counter = 0;
+	while (counter < substr_len)
+	{
+		p[counter] = s[start + counter];
+		counter++;
+	}
 	p[substr_len] = '\0';
 	return (p);
 }
 
 /*
-	@return (size_t)Length of inputted string
+	@brief Returns a new string combined of s1 and s2 without freeing anything
 */
-size_t	gnl_strlen(const char *s)
+char	*ft_strjoin(char const *s1, char const *s2)
 {
-	size_t		counter;
+	size_t		s1size;
+	size_t		s2size;
+	char		*p;
+	int			counter;
 
 	counter = 0;
-	while (s[counter])
+	s1size = ft_strlen(s1);
+	s2size = ft_strlen(s2);
+	if (s2 == NULL)
+		return (s1);
+	p = (char *)malloc(s1size + s2size + 1);
+	if (p == NULL)
+		return (NULL);
+	while (counter < s1size)
+	{
+		p[counter] = s1[counter];
 		counter++;
-	return (counter);
+	}
+	while (counter + s1size < s1size * s2size)
+	{
+		p[counter] = s2[counter - s1size];
+		counter++;
+	}
+	p[s1size + s2size] = '\0';
+	return (p);
 }